// capstone_project/Angular/e-commerce-website/src/interfaces/product.ts
export interface Product {
  id?: number;
  name: string;
  description?: string;
  price: number;
  quantity: number;
  imageUrl?: string;
}

export interface ProductDTO {
  id?: number;
  name: string;
  description: string;
  price: number;
  quantity: number;
  image?: File;
}

export interface CartItem {
  product: Product;
  quantity: number;
}